additional environment vars corresponding to CLI options
